move ColumnFilterMap to utils dir
