+++ core/src/test/scala/filodb.core/UtilsSpec.scala
@@ -1,6 +1,6 @@
 package filodb.core
 
-import filodb.core.Utils.ColumnFilterMap
+import filodb.core.utils.ColumnFilterMap
 import filodb.core.query.{ColumnFilter, Filter}
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
