+++ spark-jobs/src/main/scala/filodb/downsampler/chunk/DownsamplerSettings.scala
@@ -9,9 +9,9 @@ import net.ceedubs.ficus.Ficus._
 import org.apache.spark.sql.types._
 
 import filodb.coordinator.{FilodbSettings, NodeClusterActor}
-import filodb.core.Utils.ColumnFilterMap
 import filodb.core.query.ColumnFilter
 import filodb.core.store.{IngestionConfig, StoreConfig}
+import filodb.core.utils.ColumnFilterMap
 import filodb.downsampler.DownsamplerContext
 import filodb.downsampler.chunk.ExportConstants._
 import filodb.prometheus.ast.InstantExpression
